requirements.txt
scikit-learn
pandas
requests
numpy
mysql-connector-python
pymysql
streamlit
